backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
openai==0.28.1
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
gunicorn==21.2.0
